feat(extract_var)!: add support for c# and go back to relying on captures for output range
Computing the output range allows it to be as close as possible to the
highest extracted text at the cost of thinking about "outside" regions
of scopes. This becomes a problem in languages like Java or C# where the
outside of a function may be the inside of a class, where the semantics
for variable declarations are different.

The solution is to instead think about the "inside" of scopes and use
that as the output range, at the cost of it being as close as possible
to the highest extracted text.